reorganize java package with subpackages
