+++ DatabindingWithNewsApi/app/src/main/java/com/enpassio/databindingwithnewsapi/model/Article.java
@@ -1,4 +1,4 @@
-package com.enpassio.databindingwithnewsapi;
+package com.enpassio.databindingwithnewsapi.model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
