+++ DatabindingWithNewsApi/app/src/main/java/com/enpassio/databindingwithnewsapi/ui/ArticleListFragment.java
@@ -1,4 +1,4 @@
-package com.enpassio.databindingwithnewsapi;
+package com.enpassio.databindingwithnewsapi.ui;
 
 import android.arch.lifecycle.ViewModelProviders;
 import android.databinding.DataBindingUtil;
@@ -12,7 +12,10 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.enpassio.databindingwithnewsapi.R;
 import com.enpassio.databindingwithnewsapi.databinding.FragmentListBinding;
+import com.enpassio.databindingwithnewsapi.model.Article;
+import com.enpassio.databindingwithnewsapi.viewmodel.MainViewModel;
 
 public class ArticleListFragment extends Fragment implements NewsAdapter.ArticleClickListener {
