+++ DatabindingWithNewsApi/app/src/main/java/com/enpassio/databindingwithnewsapi/ui/MainActivity.java
@@ -1,7 +1,9 @@
-package com.enpassio.databindingwithnewsapi;
+package com.enpassio.databindingwithnewsapi.ui;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+
+import com.enpassio.databindingwithnewsapi.R;
 
 public class MainActivity extends AppCompatActivity {
