+++ DatabindingWithNewsApi/app/src/main/java/com/enpassio/databindingwithnewsapi/utils/NewsAppModule.java
@@ -1,4 +1,4 @@
-package com.enpassio.databindingwithnewsapi;
+package com.enpassio.databindingwithnewsapi.utils;
 
 import com.bumptech.glide.annotation.GlideModule;
 import com.bumptech.glide.module.AppGlideModule;
